Refactor of slack module and commands processing
